- Added var keyword to declare a variable - Moved the ConsoleOutput into a separate file - Rename the test methods
