+++ csharp-advanced-topics/ActionAndFuncDelegatesInCSharp/Test/FuncDelegateUnitTest.cs
@@ -9,7 +9,7 @@ namespace TestProject1
         [Theory]
         [InlineData(2, true)]
         [InlineData(3, false)]
-        public void IsEven_ShouldReturnCorrectResult(int number, bool expected)
+        public void WhenNumberIsPassedToIsEven_ThenShouldReturnANumberIsEvenOrNot(int number, bool expected)
         {
             // Act
             bool result = ActionFuncDelegate.IsEven(number);
